+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * 【core   】     修复createScheduledExecutor单位不是毫秒的问题（issue#I3OYIW@Gitee）
 * 【core   】     修复Tailer无stop问题（issue#I3PQLQ@Gitee）
 * 【core   】     修复空白excel读取报错问题（issue#1552@Github）
+* 【extra  】     修复Sftp.mkDirs报错问题（issue#1536@Github）
 
 -------------------------------------------------------------------------------------------------------------